requirements.txt
asttokens==2.1.0
backcall==0.2.0
bokeh==3.0.2
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.6
cycler==0.11.0
decorator==5.1.1
executing==1.2.0
Flask==2.2.2
fonttools==4.38.0
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
googleapis-common-protos==1.57.0
gunicorn==20.1.0
httplib2==0.21.0
idna==3.4
ipython==8.6.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
numpy==1.23.5
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
parso==0.8.3
pathlib==1.0.1
pickleshare==0.7.5
Pillow==9.3.0
prompt-toolkit==3.0.33
protobuf==4.21.9
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.3
seaborn==0.12.1
simplegeneric==0.8.1
six==1.16.0
stack-data==0.6.1
tabulate==0.9.0
tornado==6.2
traitlets==5.5.0
uritemplate==4.1.1
urllib3==1.26.12
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.2.2
xyzservices==2022.9.0
